+++ tests/test_02_dxf_graphics/test_212_insert.py
@@ -7,6 +7,7 @@ import ezdxf
 from ezdxf.entities.insert import Insert
 from ezdxf.lldxf.const import DXF12, DXF2000
 from ezdxf.lldxf.tagwriter import TagCollector, basic_tags_from_text
+from ezdxf.math import Vector
 
 TEST_CLASS = Insert
 TEST_TYPE = 'INSERT'
@@ -209,3 +210,58 @@ def test_copy_with_insert(doc):
     assert copy.dxf.handle != insert.dxf.handle
     assert copy.dxf.owner == psp.layout_key
     assert copy.attribs[0].dxf.owner == psp.layout_key
+
+
+def test_brcs_no_transform():
+    insert = TEST_CLASS.new(handle='ABBA', owner='0')
+    brcs = insert.brcs()
+    assert brcs.to_wcs((0, 0, 0)) == (0, 0, 0)
+    assert brcs.direction_to_wcs((1, 0, 0)) == (1, 0, 0)
+
+
+def test_brcs_insert():
+    insert = TEST_CLASS.new(handle='ABBA', owner='0', dxfattribs={
+        'insert': (1, 2, 3),
+    })
+    brcs = insert.brcs()
+    assert brcs.to_wcs((0, 0, 0)) == (1, 2, 3)
+    assert brcs.direction_to_wcs((1, 0, 0)) == (1, 0, 0)
+
+
+def test_brcs_insert_and_base_point(doc):
+    doc.blocks.new('BRCS_001', base_point=(2, 2, 2))
+    insert = doc.modelspace().add_blockref('BRCS_001', insert=(1, 2, 3))
+    brcs = insert.brcs()
+    assert brcs.to_wcs((0, 0, 0)) == (-1, 0, 1)
+    assert brcs.direction_to_wcs((1, 0, 0)) == (1, 0, 0)
+
+
+def test_brcs_rotation():
+    insert = TEST_CLASS.new(handle='ABBA', owner='0', dxfattribs={
+        'insert': (0, 0, 0),
+        'rotation': 90,
+    })
+    brcs = insert.brcs()
+    assert list(brcs.points_to_wcs([(1, 0, 0), (0, 0, 1)])) == [(0, 1, 0), (0, 0, 1)]
+    assert brcs.direction_to_wcs((1, 0, 0)) == (0, 1, 0)
+
+
+def test_brcs_scaled():
+    insert = TEST_CLASS.new(handle='ABBA', owner='0', dxfattribs={
+        'xscale': 2,
+        'yscale': 3,
+        'zscale': 4,
+    })
+    brcs = insert.brcs()
+    assert brcs.to_wcs((1, 1, 1)) == (2, 3, 4)
+    assert brcs.direction_to_wcs((1, 0, 0)) == (2, 0, 0), 'scaling has to be applied for directions'
+
+
+def test_brcs_dircection():
+    insert = TEST_CLASS.new(handle='ABBA', owner='0', dxfattribs={
+        'insert': (1, 2, 3),
+        'xscale': 2,
+    })
+    brcs = insert.brcs()
+    assert brcs.to_wcs((1, 0, 0)) == (3, 2, 3)
+    assert brcs.direction_to_wcs((1, 0, 0)) == (2, 0, 0), 'only scaling has to be applied for directions'